+++ CHANGELOG/changelog.md
@@ -4,6 +4,7 @@ Teamwire on-premise platform release August 2019
 ### Enhancements:
 - Introduce new changelog.
 - twctl tool: The connectivity function is enhanced with an ocsp check 
+- plugin: Add a new ansible plugin that compares the used platform version with the current repository version.
 
 ### Bugfixes 2019-07:
 - [2019-07.1] OCSP: Ocsp check during playbook run is not enforced anymore.